+++ src/commonMain/kotlin/org/intellij/markdown/parser/TopLevelBuilder.kt
@@ -5,21 +5,26 @@ import org.intellij.markdown.MarkdownTokenTypes
 import org.intellij.markdown.ast.ASTNode
 import org.intellij.markdown.ast.ASTNodeBuilder
 
-class TopLevelBuilder(nodeBuilder: ASTNodeBuilder) : TreeBuilder(nodeBuilder) {
+class TopLevelBuilder(
+    nodeBuilder: ASTNodeBuilder,
+    cancellationToken: CancellationToken
+) : TreeBuilder(nodeBuilder, cancellationToken) {
+    @Deprecated("Use constructor with CancellationToken")
+    constructor(nodeBuilder: ASTNodeBuilder): this(nodeBuilder, CancellationToken.NonCancellable)
 
-    override fun flushEverythingBeforeEvent(event: TreeBuilder.MyEvent, currentNodeChildren: MutableList<TreeBuilder.MyASTNodeWrapper>?) {
+    override fun flushEverythingBeforeEvent(event: MyEvent, currentNodeChildren: MutableList<MyASTNodeWrapper>?) {
     }
 
-    override fun createASTNodeOnClosingEvent(event: TreeBuilder.MyEvent, currentNodeChildren: List<TreeBuilder.MyASTNodeWrapper>, isTopmostNode: Boolean): TreeBuilder.MyASTNodeWrapper {
+    override fun createASTNodeOnClosingEvent(event: MyEvent, currentNodeChildren: List<MyASTNodeWrapper>, isTopmostNode: Boolean): MyASTNodeWrapper {
         val newNode: ASTNode
 
-        val type = event.info.`type`
+        val type = event.info.type
         val startOffset = event.info.range.first
         val endOffset = event.info.range.last
 
         if (type is MarkdownElementType && type.isToken) {
             val nodes = nodeBuilder.createLeafNodes(type, startOffset, endOffset)
-            return TreeBuilder.MyASTNodeWrapper(nodes.first(), startOffset, endOffset)
+            return MyASTNodeWrapper(nodes.first(), startOffset, endOffset)
         }
 
         val childrenWithWhitespaces = ArrayList<ASTNode>(currentNodeChildren.size)
@@ -29,7 +34,7 @@ class TopLevelBuilder(nodeBuilder: ASTNodeBuilder) : TreeBuilder(nodeBuilder) {
                     startOffset,
                     currentNodeChildren.firstOrNull()?.startTokenIndex ?: endOffset)
 
-            for (i in 1..currentNodeChildren.size - 1) {
+            for (i in 1 until currentNodeChildren.size) {
                 val prev = currentNodeChildren.get(i - 1)
                 val next = currentNodeChildren.get(i)
 
@@ -37,14 +42,14 @@ class TopLevelBuilder(nodeBuilder: ASTNodeBuilder) : TreeBuilder(nodeBuilder) {
 
                 addRawTokens(childrenWithWhitespaces, prev.endTokenIndex, next.startTokenIndex)
             }
-            if (!currentNodeChildren.isEmpty()) {
+            if (currentNodeChildren.isNotEmpty()) {
                 childrenWithWhitespaces.add(currentNodeChildren.last().astNode)
                 addRawTokens(childrenWithWhitespaces, currentNodeChildren.last().endTokenIndex, endOffset)
             }
 //        }
 
         newNode = nodeBuilder.createCompositeNode(type, childrenWithWhitespaces)
-        return TreeBuilder.MyASTNodeWrapper(newNode, startOffset, endOffset)
+        return MyASTNodeWrapper(newNode, startOffset, endOffset)
     }
 
     private fun addRawTokens(childrenWithWhitespaces: MutableList<ASTNode>, from: Int, to: Int) {